+++ Userland/Libraries/LibPDF/ColorSpace.h
@@ -8,7 +8,6 @@
 
 #include <AK/FlyString.h>
 #include <LibGfx/Color.h>
-#include <LibPDF/Object.h>
 #include <LibPDF/Value.h>
 
 #define ENUMERATE_COLOR_SPACES(V) \
